hackernews/requirements.txt
aniso8601==7.0.0
asgiref==3.7.2
Django==2.2
django-environ==0.11.2
django-filter==2.1.0
django-graphql-jwt==0.2.1
graphene==2.1.9
graphene-django==2.2.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
PyJWT==1.7.0
pytz==2024.1
Rx==1.6.3
singledispatch==4.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
